Update abstract.py
While installing adobepdf and ie11 dependencies the installation get stuck due to "wait_process_appear" function in infinitive loop. This commit change loop into limited loop. 
It is only a test. While it can indicate problem with instalation, Wait_process_appear is not necessary for successful instalation, which was verified in graphical interface.